requirements.txt
setuptools>=65.0.0,<81
face_recognition>=1.3.0
git+https://github.com/ageitgey/face_recognition_models
opencv-python>=4.8.0
dlib>=19.24.0
PyYAML>=6.0
dbus-python>=1.3.2
cryptography>=41.0.0
sqlalchemy>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
python-pam>=2.0.2
tqdm>=4.66.0
colorama>=0.4.6
